+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/hostenvironment/HostEnvironmentInfo.kt
@@ -7,5 +7,6 @@ import androidx.annotation.Keep
 data class HostEnvironmentInfo(
     val platformVersion: String,
     val hostVersion: String,
-    val sdkVersion: String
+    val sdkVersion: String,
+    val hostLocale: String
 )