+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## CHANGELOG
 
 ### 3.9.0 (2021-XX-XX)
+**SDK**
+- **Feature:** Added `hostLocale` in `HostEnvironmentInfo` to provide default language value from Host App.
+
 **Sample App**
 - **Feature:** Added production and staging toggle to change environments.